+++ CHANGELOG.md
@@ -21,6 +21,7 @@
  *   Warn before loading a file if there are unsaved changes
  *   Now grid lines are configured by number of divisions instead of pixels between divisions
  *   Basic setup of Function eases the creation of parameters
+ *   Now remembers setup type (basic/advanced) in Build view
  *   Fix: Redo block add after undo
  *   Fix: Changing mode right after loading a file showed a blank code
  *   Fix: If opening code on Touch mode show Run button